Fix date() to allow Date type input
This is triggered by the "4.17.42 " commit on 15 Sep 2015, which
standardised moment.
The change makes parseInputDate call getMoment which creates moment with
parseFormats. The presence of the format parameter makes moment treat it
as a string (before checking for a date type).
https://github.com/Eonasdan/bootstrap-datetimepicker/issues/1616

Fixed getMoment to handle Date input types

Also fixes defaultDate() & viewDate()